Consolidate save migrator types
