--- src/@types/settings-save-migrator.ts
@@ -1,6 +1,0 @@
-import type { Settings } from "#types/settings";
-
-export interface SettingsSaveMigrator {
-  version: string;
-  migrate: (data: Partial<Settings>) => void;
-}
