--- src/@types/system-save-migrator.ts
@@ -1,6 +1,0 @@
-import type { SystemSaveData } from "#types/system-data";
-
-export interface SystemSaveMigrator {
-  version: string;
-  migrate: (data: SystemSaveData) => void;
-}
